delete type file
