+++ packages/twenty-shared/src/application/frontComponentManifestType.ts
@@ -1,4 +1,13 @@
-import { type CommandMenuItemManifest } from './commandMenuItemManifestType';
+import { type SyncableEntityOptions } from '@/application/syncableEntityOptionsType';
+
+export type CommandMenuItemManifest = SyncableEntityOptions & {
+  label: string;
+  icon?: string;
+  isPinned?: boolean;
+  availabilityType?: 'GLOBAL' | 'SINGLE_RECORD' | 'BULK_RECORDS';
+  availabilityObjectUniversalIdentifier?: string;
+  frontComponentUniversalIdentifier: string;
+};
 
 export type FrontComponentCommandManifest = Omit<
   CommandMenuItemManifest,
